rename constants to positions.js
